Added a projects page to the website
Added a page to the website that contains a list of the various projects that I have worked on.

Added an entry to the projects page for Jekyll Helper. Also added a screenshot of Jekyll Helper.

Also added a rule to the website's css so that it properly bolds strong text.

Closes #26